Always show last analyzed time point popup (#1408)
