+++ frontend/src/components/MapView/Layers/AnticipatoryActionStormLayer/AAStormLandfallPopup/PopupContent/index.test.tsx
@@ -1,5 +1,6 @@
 import { render } from '@testing-library/react';
-import { AACategory } from 'context/anticipatoryAction/AAStormStateSlice/types';
+
+import { AACategory } from 'context/anticipatoryAction/AAStormStateSlice/parsedStormDataTypes';
 import PopupContent from '.';
 
 describe('AAStormLandfallPopup component', () => {
