+++ frontend/src/components/MapView/LeftPanel/AnticipatoryActionPanel/AnticipatoryActionStormPanel/ActivationTriggerView/types.ts
@@ -1,0 +1,13 @@
+import { AACategory } from 'context/anticipatoryAction/AAStormStateSlice/parsedStormDataTypes';
+
+export const AAPanelCategories: AACategory[] = [
+  AACategory.Severe,
+  AACategory.Moderate,
+];
+
+export const AADisplayCategory: {
+  [key in AACategory]?: string;
+} = {
+  [AACategory.Severe]: ' >118 KM/H',
+  [AACategory.Moderate]: ' >89 KM/H',
+};
