+++ frontend/src/context/anticipatoryAction/AAStormStateSlice/index.ts
@@ -1,13 +1,10 @@
 import { createAsyncThunk, createSlice } from '@reduxjs/toolkit';
 import { DateItem } from 'config/types';
 import type { CreateAsyncThunkTypes, RootState } from '../../store';
-import {
-  AAStormData,
-  AAStormWindStateReports,
-  AnticipatoryActionState,
-  StormData,
-} from './types';
+import { AAStormWindStateReports, AnticipatoryActionState } from './types';
 import { parseAndTransformAA } from './utils';
+import { ParsedStormData } from './parsedStormDataTypes';
+import { StormDataResponseBody } from './rawStormDataTypes';
 
 const initialState: AnticipatoryActionState = {
   data: {},
@@ -28,7 +25,7 @@ export const loadAllAAStormData = createAsyncThunk<
 
 export const loadLatestStormReport = createAsyncThunk<
   {
-    data: AAStormData;
+    data: ParsedStormData;
   },
   undefined,
   CreateAsyncThunkTypes
@@ -40,7 +37,7 @@ export const loadLatestStormReport = createAsyncThunk<
         'https://data.earthobservation.vam.wfp.org/public-share/aa/ts/outputs/latest.json',
       );
       const stormData = await response.json();
-      const data = parseAndTransformAA(stormData as StormData);
+      const data = parseAndTransformAA(stormData as StormDataResponseBody);
       return data;
     } catch (error) {
       return rejectWithValue(error);
@@ -50,7 +47,7 @@ export const loadLatestStormReport = createAsyncThunk<
 
 export const loadStormReport = createAsyncThunk<
   {
-    data: AAStormData;
+    data: ParsedStormData;
   },
   { stormName: string; date: string },
   CreateAsyncThunkTypes
@@ -67,7 +64,7 @@ export const loadStormReport = createAsyncThunk<
         `https://data.earthobservation.vam.wfp.org/public-share/aa/ts/outputs/${stormName}/${date}.json?v2`,
       );
       const stormData = await response.json();
-      const data = parseAndTransformAA(stormData as StormData);
+      const data = parseAndTransformAA(stormData as StormDataResponseBody);
       return data;
     } catch (error) {
       return rejectWithValue(error);
